Converted scenario-lang to single machine step
I added a end-of-commands symbol to a thread in the syntax. I also
modified a trace-entry to always be a pair. A trace-entry that does
not need to specify an action ID uses the end-of-commands symbol.

I expanded the expression state to indicate the action ID of a send if
one is specified in the trace. I modified all of the expression rules
to be sensitive to this extra state member. In particular, I changed
the "Recvi Wait Cmd 0" to grab the value of the extra state member to
create the (wait aid_r aid_s) pair needed for the following
reductions.

I updated all of the tests to the new form of the machine and
expression states. All of the tests pass as before.

More details on the conversion to modify traces to be (ploc ⊥) or
(ploc aid):

  * Add new state entry in expression machine that is (⊥ or aid) --
    done
  * Rewrite machine rules to one rule on the new expression state --
    done
  * Rewrite expression rules to see the new member of expression state
    -- done
  * Rewrite the wait rules to distinguish on ⊥ for wait-send and aid
    for wait-receive
      - Simplify to have a single "wait" command in grammar -- done
        (sortof)
      - Use the extra state member to track through the lookups and
        final match -- may not need