Reduced to single wait for send/recv and added last
There is now a single wait for send commands and a single wait for
receive commands. I have also added in a "last" structure that
enables the addition of a HB relation between the "last" send on
a given endpoint. I believe "last" can be removed from the state
as program order gives the same guarnatee.