Modified trace to include queue movement
The trace has been extensively modified to drop the action ID on
a match for a wait and to add a list of optional queue movement.
The match pair is determined by queue movement and not a specified
action ID. To accommodate queue movement, we added a new machine
to process the queue actions. We also added a set of queues for
each end-point.

Intuitively, pending_s is the pending sends. Pending_r is the
pending receives. Q-set is the delivered messages. Queue
movement transfers a pending send to a delievered send (i.e.,
from pending_s to q-set).